fix: rename to more generic `rotation.ts`
